+++ packages/visual-stack-docs/src/containers/Components/Docs/expanding-input-button.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Body, Header, Panel } from '@cjdev/visual-stack/lib/components/Panel';
-import { Demo, Snippet } from '../../../components/Demo';
+import { Demo, MarginTopSnippet as Snippet } from '../../../components/Demo';
 import { ExpandingInputButton } from '@cjdev/visual-stack/lib/components/ExpandingInputButton';
 import MagnifyIcon from 'mdi-react/MagnifyIcon';