+++ packages/visual-stack-docs/src/containers/Components/Docs/index.js
@@ -5,33 +5,34 @@ import {
 } from '@cjdev/visual-stack/lib/components/PageHeader';
 import PageContent from '@cjdev/visual-stack/lib/components/PageContent';
 // 1. add an import for your demo
-import BlankSlateDocs from './blankslate';
-import ButtonWithDropdownDocs from './button-with-dropdown';
-import ButtonDocs from './button';
 import AlertDocs from './alert';
+import BadgeDocs from './badge';
+import BlankSlateDocs from './blankslate';
+import BoxDocs from './box';
+import ButtonDocs from './button';
+import ButtonWithDropdownDocs from './button-with-dropdown';
+import CardDocs from './card';
+import CollapsiblePanelDocs from './collapsiblepanel';
 import DatePickerDocs from './datepicker';
+import DrawerDocs from './drawer';
+import ExpandingInputButtonDocs from './expanding-input-button';
 import FormDocs from './form';
 import ListDocs from './list';
+import ListViewDocs from './listview';
+import LoadingAnimationDocs from './loading-animation';
 import ModalDocs from './modal';
-import PanelDocs from './panel';
 import PageHeaderDocs from './pageheader';
+import PaginationDocs from './pagination';
+import PanelDocs from './panel';
+import PercentSliderDocs from './percentslider';
+import PopoverDocs from './popover';
 import SelectDocs from './select';
 import SideNavDocs from './sidenav';
 import SlidingPanelDocs from './slidingpanel';
 import SpinnerDocs from './spinner';
-import TableDocs from './table';
 import TabLayoutDocs from './tablayout';
-import CardDocs from './card';
-import BadgeDocs from './badge';
-import ListViewDocs from './listview';
-import ExpandingInputButtonDocs from './expanding-input-button';
-import LoadingAnimationDocs from './loading-animation';
-import PopoverDocs from './popover';
-
-import CollapsiblePanelDocs from './collapsiblepanel';
-import PaginationDocs from './pagination';
-import PercentSliderDocs from './percentslider';
-import TopNavDocs from "./topnav";
+import TableDocs from './table';
+import TopNavDocs from './topnav';
 
 const routeComponentMap = {};
 
@@ -40,53 +41,50 @@ const addComponentRoute = (path, linkName, component) => {
 };
 
 // 2. add your demo to the routeComponentMap
+addComponentRoute('alert', 'Alert', <AlertDocs />);
+addComponentRoute('badge', 'Badge', <BadgeDocs />);
 addComponentRoute('blankslate', 'Blank Slate', <BlankSlateDocs />);
+addComponentRoute('box', 'Box', <BoxDocs />);
 addComponentRoute('button', 'Button', <ButtonDocs />);
 addComponentRoute(
   'button-with-dropdown',
   'Button With Dropdown',
   <ButtonWithDropdownDocs />
 );
-addComponentRoute('alert', 'Alert', <AlertDocs />);
-addComponentRoute('form', 'Form', <FormDocs />);
-addComponentRoute('list', 'List', <ListDocs />);
-addComponentRoute('modal', 'Modal', <ModalDocs />);
-addComponentRoute('panel', 'Panel', <PanelDocs />);
-addComponentRoute('pageheader', 'Page Header', <PageHeaderDocs />);
-addComponentRoute('percentslider', 'Percent Slider', <PercentSliderDocs />);
-addComponentRoute('select', 'Select', <SelectDocs />);
-addComponentRoute('sidenav', 'SideNav', <SideNavDocs />);
-addComponentRoute('topnav', 'TopNav', <TopNavDocs />);
-addComponentRoute('slidingpanel', 'Sliding Panel', <SlidingPanelDocs />);
-addComponentRoute(
-  'loading-animation',
-  'Loading Animation',
-  <LoadingAnimationDocs />
-);
-addComponentRoute('popover', 'Popover', <PopoverDocs />);
-addComponentRoute('spinner', 'Spinner', <SpinnerDocs />);
-addComponentRoute('table', 'Table', <TableDocs />);
-addComponentRoute('tablayout', 'TabLayout', <TabLayoutDocs />);
 addComponentRoute('card', 'Card', <CardDocs />);
-addComponentRoute('badge', 'Badge', <BadgeDocs />);
-addComponentRoute('listview', 'List View', <ListViewDocs />);
-addComponentRoute('pagination', 'Pagination', <PaginationDocs />);
-addComponentRoute(
-  'expanding-input-button',
-  'Expanding Input Button',
-  <ExpandingInputButtonDocs />
-);
 addComponentRoute(
   'collapsiblepanel',
   'Collapsible Panel',
   <CollapsiblePanelDocs />
 );
 addComponentRoute(
-  'button-with-dropdown',
-  'Button With Dropdown',
-  <ButtonWithDropdownDocs />
+  'expanding-input-button',
+  'Expanding Input Button',
+  <ExpandingInputButtonDocs />
 );
 addComponentRoute('datepicker', 'Date Picker', <DatePickerDocs />);
+addComponentRoute('drawer', 'Drawer', <DrawerDocs />);
+addComponentRoute('form', 'Form', <FormDocs />);
+addComponentRoute('list', 'List', <ListDocs />);
+addComponentRoute('listview', 'List View', <ListViewDocs />);
+addComponentRoute(
+  'loading-animation',
+  'Loading Animation',
+  <LoadingAnimationDocs />
+);
+addComponentRoute('modal', 'Modal', <ModalDocs />);
+addComponentRoute('pageheader', 'Page Header', <PageHeaderDocs />);
+addComponentRoute('pagination', 'Pagination', <PaginationDocs />);
+addComponentRoute('panel', 'Panel', <PanelDocs />);
+addComponentRoute('percentslider', 'Percent Slider', <PercentSliderDocs />);
+addComponentRoute('popover', 'Popover', <PopoverDocs />);
+addComponentRoute('select', 'Select', <SelectDocs />);
+addComponentRoute('sidenav', 'SideNav', <SideNavDocs />);
+addComponentRoute('slidingpanel', 'Sliding Panel', <SlidingPanelDocs />);
+addComponentRoute('spinner', 'Spinner', <SpinnerDocs />);
+addComponentRoute('tablayout', 'TabLayout', <TabLayoutDocs />);
+addComponentRoute('table', 'Table', <TableDocs />);
+addComponentRoute('topnav', 'TopNav', <TopNavDocs />);
 
 const ComponentDocs = ({ params }) => {
   const routeData = routeComponentMap[params.componentName];