+++ packages/visual-stack-docs/src/containers/Components/Docs/pagination.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-invalid-this */
 import React from 'react';
 import { Panel, Body, Header } from '@cjdev/visual-stack/lib/components/Panel';
-import { Demo, Snippet } from '../../../components/Demo';
+import { Demo, MarginTopSnippet as Snippet } from '../../../components/Demo';
 
 /* s1:start */
 import Pagination from '@cjdev/visual-stack-redux/lib/components/Pagination';