+++ packages/visual-stack-docs/src/containers/Components/Docs/select.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-console */
 import React from 'react';
-import { Demo, Snippet } from '../../../components/Demo';
+import { Demo, MarginTopSnippet as Snippet } from '../../../components/Demo';
 import { Body, Header, Panel } from '@cjdev/visual-stack/lib/components/Panel';
 
 /* s6:start */