+++ packages/visual-stack-docs/src/containers/Components/Docs/table.js
@@ -2,7 +2,7 @@
 /** @prettier */
 import React, { Fragment } from 'react';
 import { Panel, Body, Header } from '@cjdev/visual-stack/lib/components/Panel';
-import { Demo, Snippet } from '../../../components/Demo';
+import { Demo, MarginTopSnippet as Snippet } from '../../../components/Demo';
 
 /* s1:start */
 import {