+++ packages/visual-stack-docs/src/containers/Experimental/Docs/index.js
@@ -7,7 +7,6 @@ import PageContent from '@cjdev/visual-stack/lib/components/PageContent';
 
 // 1. add an import for your demo
 import TextDocs from './text';
-import BoxDocs from './box';
 import Tree from './tree';
 import TriStateCheckboxDocs from './tristatecheckbox';
 
@@ -19,7 +18,6 @@ const addComponentRoute = (path, linkName, component) => {
 
 // 2. add your demo to the experimentalRouteMap
 addComponentRoute('text', 'Text', <TextDocs />);
-addComponentRoute('box', 'Box', <BoxDocs />);
 addComponentRoute('tree', 'Tree', <Tree />);
 addComponentRoute(
   'tristatecheckbox',