+++ packages/visual-stack/src/components/Box/index.js
@@ -1,17 +1,16 @@
-/* eslint-disable no-console */
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
 import './Box.css';
 
-const Box = ({
+export const Box = ({
   align,
   border,
   children,
   className,
   direction,
   gap,
-  grow,
+  expand,
   justify,
   padding,
   paddingBottom,
@@ -34,7 +33,7 @@ const Box = ({
       paddingTop && `vs-box-padding-top-${paddingTop}`,
       align && `vs-box-align-${align}`,
       justify && `vs-box-justify-${justify}`,
-      grow && 'vs-box-grow',
+      expand && 'vs-box-expand',
       className
     )}
   >
@@ -43,7 +42,7 @@ const Box = ({
 );
 
 Box.defaultProps = {
-  direction: 'row',
+  direction: 'column',
 };
 
 Box.propTypes = {
@@ -52,7 +51,7 @@ Box.propTypes = {
   direction: PropTypes.oneOf(['column', 'row']),
   padding: PropTypes.oneOf([
     'small',
-    'default',
+    'medium',
     'large',
     'xl',
     '2xl',
@@ -61,7 +60,7 @@ Box.propTypes = {
   ]),
   paddingTop: PropTypes.oneOf([
     'small',
-    'default',
+    'medium',
     'large',
     'xl',
     '2xl',
@@ -70,7 +69,7 @@ Box.propTypes = {
   ]),
   paddingBottom: PropTypes.oneOf([
     'small',
-    'default',
+    'medium',
     'large',
     'xl',
     '2xl',
@@ -79,7 +78,7 @@ Box.propTypes = {
   ]),
   paddingLeft: PropTypes.oneOf([
     'small',
-    'default',
+    'medium',
     'large',
     'xl',
     '2xl',
@@ -88,22 +87,14 @@ Box.propTypes = {
   ]),
   paddingRight: PropTypes.oneOf([
     'small',
-    'default',
+    'medium',
     'large',
     'xl',
     '2xl',
     '3xl',
     '4xl',
   ]),
-  gap: PropTypes.oneOf([
-    'small',
-    'default',
-    'large',
-    'xl',
-    '2xl',
-    '3xl',
-    '4xl',
-  ]),
+  gap: PropTypes.oneOf(['small', 'medium', 'large', 'xl', '2xl', '3xl', '4xl']),
   align: PropTypes.oneOf(['start', 'center', 'end']),
   justify: PropTypes.oneOf([
     'start',
@@ -114,7 +105,5 @@ Box.propTypes = {
     'space-evenly',
   ]),
   border: PropTypes.bool,
-  grow: PropTypes.bool,
+  expand: PropTypes.bool,
 };
-
-export default Box;