+++ packages/visual-stack/src/components/TopNav/BackToCjLink.js
@@ -1,17 +1,12 @@
-import React from 'react'
+import React from 'react';
 
 import Box from '../../experimental/Box';
-import Text from '../../experimental/Text'
-import ArrowLeftIcon from 'mdi-react/ArrowLeftIcon'
-import PropTypes from 'prop-types'
-import cn from 'classnames'
+import Text from '../../experimental/Text';
+import ArrowLeftIcon from 'mdi-react/ArrowLeftIcon';
+import PropTypes from 'prop-types';
+import cn from 'classnames';
 
-export const BackToCjLink = ({
-                    className,
-                    title,
-                    onClick,
-                    ...restProps
-}) => {
+export const BackToCjLink = ({ className, title, onClick, ...restProps }) => {
   return (
     <Box
       {...restProps}
@@ -22,13 +17,17 @@ export const BackToCjLink = ({
       onClick={onClick}
       className="topnav-action"
     >
-      <Text id="topnav-back-icon" className="vs-topnav-action-text"><ArrowLeftIcon /></Text>
-      <Text id="topnav-back-title" className="vs-topnav-action-text">{title}</Text>
+      <Text id="topnav-back-icon" className="vs-topnav-action-text">
+        <ArrowLeftIcon />
+      </Text>
+      <Text id="topnav-back-title" className="vs-topnav-action-text">
+        {title}
+      </Text>
     </Box>
-  )
-}
+  );
+};
 
 BackToCjLink.propTypes = {
   title: PropTypes.string,
   onClick: PropTypes.func,
-}
+};