+++ packages/visual-stack/src/index.js
@@ -2,8 +2,10 @@ import './global';
 
 // modules that do not need qualification. i.e., the components
 // they export can be scoped to VisualStack.<component>
+export * from './components/Box';
 export * from './components/Button';
 export * from './components/Badge';
+export * from './components/Drawer';
 export * from './components/Spinner';
 
 // modules that need scoping. i.e. the components they