+++ CHANGELOG.md
@@ -1,8 +1,49 @@
+# 7.3.2
 
+## Fix
+
+- DialogLayout: Fixed the padding at the top of the dialog layout.
+
+# 7.3.1
+
+## Fix
+
+- Drawer: Added missing propTypes
+
+# 7.3.0
+
+## New Component
+
+- Add Drawer component
+
+# 7.2.1
+
+## Fix
+
+- Box: remove console warning on experimental Box
+
+# 7.2.0
+
+## Updates
+
+- Box: move out of experimental. Experimental version is kept for backward compatibility but deprecated. Will be removed in a future release.
+- Box: rename spacing props values from "default" to "medium".
+- Box: update Box docs.
+
+# 7.1.3
+
+- Update to CONTRIBUTING.md about installing lerna.
+- BlankSlate: Remove hard-coded / default icon graphic and white background Panel per UX.
+- Form: Remove 24px margin between fields per UX.
+- Modal: Remove header and footer boarder lines, push top down, add left & right padding of 25px, include a -md class which is not quite as wide as the default. All per UX.
+- DataTable: Change the thickness and color of the header row bottom border, remove the extra DataTable.css import which was causing issues, added alignment argument to header props. All per UX.
+- TopNav: Fixed some sizing and cursor issues.
+- Panel: Now passing down classnames onto the component.
 
 # 7.1.2
 
 ## New Component
+
 - Add TopNav component
 
 # 7.1.1